move copy node operators to operators folder
